+++ CMakeLists.txt
@@ -132,9 +132,9 @@ set(kep3_SRC_FILES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/epoch.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/planet.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/lambert_problem.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/planets/keplerian.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/planets/jpl_lp.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/planets/vsop2013.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/udpla/keplerian.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/udpla/jpl_lp.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/udpla/vsop2013.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/core_astro/ic2par2ic.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/core_astro/ic2eq2ic.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/core_astro/eq2par2eq.cpp"
